Added POST request endpoint for sending emails with nodemailer
